+++ bot/models/transaction.py
@@ -1,6 +1,6 @@
 """Модель транзакций для хранения информации о покупках и продажах товаров пользователями."""
 
-from sqlalchemy import Column, Integer, String, Float, DateTime, ForeignKey
+from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, BigInteger
 
 from datetime import datetime
 
@@ -13,7 +13,7 @@ class Transaction(Base):
     __tablename__ = "transactions"
 
     id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, ForeignKey("users.telegram_id"), nullable=False)
+    user_id = Column(BigInteger, ForeignKey("users.telegram_id"), nullable=False)
     amount = Column(Integer, nullable=False)
     balance_after = Column(Integer, nullable=False)
     reason = Column(String, nullable=False)
